Helper macro for Morph impls
